requirements.txt
#pip install -r requirements.txt
aiofiles          == 23.2.1
aiogram           == 3.12.0
aiohappyeyeballs  == 2.4.0
aiohttp           == 3.10.5
aiosignal         == 1.3.1
aiosqlite         == 0.20.0
annotated-types   == 0.7.0
anyio             == 4.4.0
attrs             == 24.2.0
cachetools        == 5.5.0
certifi           == 2024.8.30
dnspython         == 2.6.1
fastapi           == 0.114.0
frozenlist        == 1.4.1
idna              == 3.8
magic-filter      == 1.0.12
motor             == 3.5.1
multidict         == 6.0.5
numpy             == 2.1.1
pip               == 24.2
pydantic          == 2.8.2
pydantic_core     == 2.20.1
pydantic-settings == 2.5.0
pymongo           == 4.8.0
python-decouple   == 3.8
python-dotenv     == 1.0.1
redis             == 5.0.8
sniffio           == 1.3.1
starlette         == 0.38.5
typing_extensions == 4.12.2
yarl              == 1.9.7
requests          == 2.32.3 
urllib3           == 2.2.2
beautifulsoup4
bs4
flet              == 0.24.1